# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FoxRunner API",
    "description": "API de pilotage du scheduler et des scenarios FoxRunner.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/graph/subscriptions": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Create Subscription",
        "operationId": "create_subscription_api_v1_graph_subscriptions_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphSubscriptionPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphSubscriptionResponsePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Graph Subscriptions Endpoint",
        "operationId": "graph_subscriptions_endpoint_api_v1_graph_subscriptions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphSubscriptionPagePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/subscriptions/{subscription_id}": {
      "patch": {
        "tags": [
          "graph"
        ],
        "summary": "Renew Subscription Endpoint",
        "operationId": "renew_subscription_endpoint_api_v1_graph_subscriptions__subscription_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subscription Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphRenewPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphSubscriptionResponsePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "graph"
        ],
        "summary": "Delete Subscription Endpoint",
        "operationId": "delete_subscription_endpoint_api_v1_graph_subscriptions__subscription_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/notifications": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Graph Notifications Endpoint",
        "operationId": "graph_notifications_endpoint_api_v1_graph_notifications_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "subscription_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphNotificationPagePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/webhook": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Graph Webhook",
        "operationId": "graph_webhook_api_v1_graph_webhook_post",
        "parameters": [
          {
            "name": "validationToken",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Validationtoken"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/lifecycle": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Graph Lifecycle Webhook",
        "operationId": "graph_lifecycle_webhook_api_v1_graph_lifecycle_post",
        "parameters": [
          {
            "name": "validationToken",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Validationtoken"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptedPayload": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          }
        },
        "type": "object",
        "required": [
          "accepted"
        ],
        "title": "AcceptedPayload"
      },
...
      "DeletedPayload": {
        "properties": {
          "deleted": {
            "type": "string",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "DeletedPayload"
      },
...
      "GraphNotificationPagePayload": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GraphNotificationPayload"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "GraphNotificationPagePayload"
      },
      "GraphNotificationPayload": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "change_type": {
            "type": "string",
            "title": "Change Type"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "client_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client State"
          },
          "lifecycle_event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Event"
          },
          "raw_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Payload"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subscription_id",
          "change_type",
          "resource",
          "raw_payload"
        ],
        "title": "GraphNotificationPayload"
      },
      "GraphRenewPayload": {
        "properties": {
          "expiration_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Expiration Datetime"
          }
        },
        "type": "object",
        "required": [
          "expiration_datetime"
        ],
        "title": "GraphRenewPayload"
      },
      "GraphSubscriptionPagePayload": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GraphSubscriptionResponsePayload"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "GraphSubscriptionPagePayload"
      },
      "GraphSubscriptionPayload": {
        "properties": {
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "change_type": {
            "type": "string",
            "title": "Change Type",
            "default": "created,updated"
          },
          "notification_url": {
            "type": "string",
            "title": "Notification Url"
          },
          "expiration_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Expiration Datetime"
          },
          "lifecycle_notification_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Notification Url"
          }
        },
        "type": "object",
        "required": [
          "resource",
          "notification_url",
          "expiration_datetime"
        ],
        "title": "GraphSubscriptionPayload"
      },
      "GraphSubscriptionResponsePayload": {
        "properties": {
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "change_type": {
            "type": "string",
            "title": "Change Type"
          },
          "notification_url": {
            "type": "string",
            "title": "Notification Url"
          },
          "lifecycle_notification_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Notification Url"
          },
          "expiration_datetime": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiration Datetime"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "subscription_id",
          "resource",
          "change_type",
          "notification_url"
        ],
        "title": "GraphSubscriptionResponsePayload"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  },
...
}